Fixes Implicit conversion loses integer precision: 'long' to 'int'
This method return NSInteger,
so change the type at operation to NSInteger.

fixes #1

+++ SPLockScreen/SPLockScreen.m
@@ -109,7 +109,7 @@
 					self.selectedCell = cell;
 				}
 				
-				int row = view.tag/kTagIdentifier - kSeed;
+				NSInteger row = view.tag/kTagIdentifier - kSeed;
 				int column = view.tag % kTagIdentifier - kSeed;
 				return row * 3 + column;
 			}
